Changed background colors from Int to Long

+++ app/src/main/java/com/wreckingballsoftware/valuewatch/data/models/BackgroundColor.kt
@@ -6,6 +6,6 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class BackgroundColor(
     val colorText: String = "White",
-    val backgroundColor: Int = 0xFFFFFFFF.toInt(),
-    val textColor: Int = 0xFF000000.toInt(),
+    val backgroundColor: Long = 0xFFFFFFFF,
+    val textColor: Long = 0xFF000000,
 ) : Parcelable
